move everything over, remove libs/api dependency from backend, --no-verify
