+++ apps/admin/backend/src/util/write_in_adjudication_table.ts
