+++ apps/admin/backend/tsconfig.build.json
@@ -10,7 +10,6 @@
     "declarationMap": true
   },
   "references": [
-    { "path": "../../../libs/api/tsconfig.build.json" },
     { "path": "../../../libs/auth/tsconfig.build.json" },
     { "path": "../../../libs/backend/tsconfig.build.json" },
     { "path": "../../../libs/basics/tsconfig.build.json" },
