+++ libs/api/src/index.ts
@@ -1,4 +1,3 @@
 export * from './base';
 export * from './client_utils';
-export * as Admin from './services/admin';
 export * as Scan from './services/scan';
