--- libs/api/src/services/admin/index.ts
@@ -1,2 +1,0 @@
-export * from './types';
-export * as Views from './views';
