--- libs/api/src/services/admin/views/index.ts
@@ -1,1 +1,0 @@
-export * as writeInAdjudicationTable from './write_in_adjudication_table';
